split the script in diferent folders
